기상예보를 보여주는 ForecastViewController 생성 및 기본 동작 구현

+++ Model/City.swift
@@ -12,7 +12,7 @@ struct SupportingCityList: Codable {
 }
 
 struct City: Codable {
-    let id: Double
+    let id: Int
     let name: String
     let state: String
     let country: String
